requirements.txt
mysql-connector-python
requests 
pandas

